# Api/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Psi Pass",
    "description": "",
    "contact": {
      "name": "Celio Vitorino Junior",
      "email": "celio.vitorino159@gmail.com"
    },
    "version": "Alpha v0.0.2-S"
  },
  "paths": {
    "/api/User/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "GetWithId",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User": {
      "get": {
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UserDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/crp/{crp}": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "crp",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/{Id}": {
      "put": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UserDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/UserSpecialty": {
      "get": {
        "tags": [
          "UserSpecialty"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/UserType": {
      "get": {
        "tags": [
          "UserType"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "UserDto": {
      "required": [
        "crp",
        "email",
        "name",
        "userSpecialtyId",
        "userTypeId"
      ],
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "format": "email",
          "type": "string"
        },
        "crp": {
          "maxLength": 6,
          "type": "string"
        },
        "userTypeId": {
          "format": "int32",
          "type": "integer"
        },
        "userSpecialtyId": {
          "format": "int32",
          "type": "integer"
        }
      }
    }
  }
}